requirements.txt
# LLM & AI
google-genai>=1.0.0

# Vector Database
chromadb>=0.4.0

# Database
mysql-connector-python>=8.0.0

# HTTP Requests (for Weather API)
requests>=2.31.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
python-dotenv>=1.0.0
